refactor: add the Kitsu:Auth module
Add the auth module for communication with Kitsu, this currently still stores the auth token inside
the Kitsu module, this should be refactored at a later point
