+++ libraries/kitsu/auth/build.gradle
@@ -18,7 +18,7 @@ android {
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
     compileOptions {
@@ -26,7 +26,7 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions {
-        jvmTarget = "1.8"
+        jvmTarget = '1.8'
     }
 }
 
@@ -34,22 +34,19 @@ dependencies {
     ext {
         retrofit_version = "2.9.0"
     }
-    implementation fileTree(dir: "libs", include: ["*.jar"])
 
     implementation project(path: ":libraries:core")
     implementation project(path: ":libraries:encryption")
-    implementation project(path: ":libraries:server")
-    implementation project(path: ":libraries:server:auth")
+    implementation project(path: ':libraries:kitsu')
+    implementation project(path: ':libraries:server:auth')
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation "androidx.appcompat:appcompat:$appcompat_version"
     implementation "androidx.core:core-ktx:$corektx_version"
     implementation "com.chesire.lintrules:lint-gradle:$lintrules_version"
     implementation "com.chesire.lintrules:lint-xml:$lintrules_version"
     implementation "com.google.dagger:hilt-android:$hilt_version"
     implementation "com.jakewharton.timber:timber:$timber_version"
-    api "com.squareup.retrofit2:converter-moshi:$retrofit_version"
-    api "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation "com.squareup.retrofit2:converter-moshi:$retrofit_version"
+    implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
 
@@ -57,6 +54,5 @@ dependencies {
     testImplementation "io.mockk:mockk:1.10.2"
     testImplementation "junit:junit:4.13.1"
 
-    kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
     kapt "com.squareup.moshi:moshi-kotlin-codegen:$moshi_version"
 }
