+++ libraries/kitsu/auth/src/main/java/com/chesire/nekome/kitsu/auth/KitsuAuthEntityMapper.kt
@@ -1,0 +1,23 @@
+package com.chesire.nekome.kitsu.auth
+
+import com.chesire.nekome.auth.api.AuthEntity
+import com.chesire.nekome.auth.api.AuthEntityMapper
+import javax.inject.Inject
+
+// TODO #387
+class KitsuAuthEntityMapper @Inject constructor() : AuthEntityMapper<KitsuAuthEntity> {
+    override fun mapFromAuthEntity(entity: AuthEntity): KitsuAuthEntity {
+        return KitsuAuthEntity(
+            "",
+            0L,
+            0L,
+            "",
+            "",
+            ""
+        )
+    }
+
+    override fun mapToAuthEntity(item: KitsuAuthEntity): AuthEntity {
+        return AuthEntity()
+    }
+}
