+++ libraries/kitsu/auth/src/main/java/com/chesire/nekome/kitsu/auth/request/LoginRequest.kt
@@ -1,16 +1,10 @@
-package com.chesire.nekome.kitsu.api.auth
+package com.chesire.nekome.kitsu.auth.request
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
-/**
- * Request object for logging in.
- */
 @JsonClass(generateAdapter = true)
 data class LoginRequest(
-    /**
-     * Username for the user, although the api expects the users email address.
-     */
     @Json(name = "username")
     val username: String,
     @Json(name = "password")
