+++ libraries/kitsu/src/main/java/com/chesire/nekome/kitsu/interceptors/AuthRefreshInterceptor.kt
@@ -1,8 +1,8 @@
 package com.chesire.nekome.kitsu.interceptors
 
-import com.chesire.nekome.kitsu.AuthProvider
+import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
-import com.chesire.nekome.server.api.AuthApi
+import com.chesire.nekome.kitsu.AuthProvider
 import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.Response
