+++ libraries/server/auth/src/main/java/com/chesire/nekome/auth/api/AuthApi.kt
@@ -4,9 +4,9 @@ import com.chesire.nekome.core.Resource
 
 interface AuthApi {
 
-    suspend fun loginAsync(username: String, password: String): Resource<Any>
+    suspend fun login(username: String, password: String): Resource<Any>
 
-    suspend fun refreshAsync(): Resource<Any>
+    suspend fun refresh(): Resource<Any>
 
-    suspend fun clearAuthAsync()
+    suspend fun clearAuth()
 }
