+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':libraries:kitsu:auth'
 include ':libraries:server:auth'
 include ':libraries:kitsu:search'
 include ':libraries:server:search'
